reorganize code and tests
